requirements.txt
Django==3.1.7
djangorestframework==3.12.4
requests==2.25.1
psycopg2>=2.8.6,<2.9
drf-spectacular>=0.15.1,<0.16
djangorestframework_simplejwt== 4.7.0
Faker>=13.0,<14.0
pandas>=1.2.0
openpyxl>=3.0.0
reportlab>=3.5.0
xlsxwriter>=1.3.0
